Turn resolvers/Query.ts file into directory
