+++ src/types/resolvers/Query/index.ts
@@ -1,0 +1,9 @@
+import { feed, filterPosts, post } from './Post';
+import { me } from './User';
+
+export const Query = [
+  me,
+  feed,
+  filterPosts,
+  post,
+];
